+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - [ ] Add missing descriptions to attributes
 - [ ] More validation rules
+- [x] Validation for `<style>` elements
+- [x] Only show Intellisense when `<html amp>` is present
 
 Features that have a checkmark are complete and available for
 download in the